Enhance types package with value based types
