+++ auth/src/index.ts
@@ -1,5 +1,6 @@
+export { isMinimumRole } from '@hike/types';
 export type { CompanyPermission } from '@hike/types';
-export { Constants, isMinimumRole, selectPreferredLocale } from '@hike/utils';
+export { Constants, selectPreferredLocale } from '@hike/utils';
 export * from './api/fetchSessionUser';
 export * from './api/trackNotification';
 export * from './errors/AuthError';
