+++ sdk/src/index.ts
@@ -1,4 +1,4 @@
 // Export functionality from SDK for external use
 export * from '@hike/services';
-export type * from '@hike/types'; // eslint-disable-line import/export
+export * from '@hike/types';
 export * from '@hike/utils';
