+++ types/src/guards/isBoolean.ts
