+++ types/src/guards/isFieldPrint.ts
@@ -1,4 +1,4 @@
-import type { FieldPrint } from '@hike/types';
+import { FieldPrint } from '../forms/FieldPrint';
 
 export const isFieldPrint = (value: unknown): value is FieldPrint =>
   value != null && typeof value === 'object' && 'position' in value && typeof value.position === 'number';
