+++ types/src/guards/isFulfilled.ts
