+++ types/src/guards/isNumber.ts
