+++ types/src/guards/isPlainObject.ts
