+++ types/src/guards/isString.ts
