+++ types/src/guards/isTruthy.ts
