+++ types/src/index.ts
@@ -1,192 +1,205 @@
 export type * from '../prisma';
-export type * from './auth/AuthSession';
-export type * from './auth/AuthStatus';
-export type * from './auth/AuthUser';
-export type * from './config/AppId';
-export type * from './config/HikeConfig';
-export type * from './config/HikeEnvironment';
-export type * from './dto/analytics/HourlyOptions';
-export type * from './dto/appointment/GenerateAppointmentsUploadLinkParams';
-export type * from './dto/appointment/ImportAppointmentsParams';
-export type * from './dto/appointment/ImportAppointmentsResponse';
-export type * from './dto/appointment/ImportPrescriptionsResponse';
-export type * from './dto/asset/GetAssetsParams';
-export type * from './dto/augment/AssetAugmentData';
-export type * from './dto/augment/AssetAugmentMediaData';
-export type * from './dto/augment/AssetAugmentMediaResult';
-export type * from './dto/augment/AssetAugmentResult';
-export type * from './dto/augment/AssetAugmentStatus';
-export type * from './dto/augment/AssetAugmentStatusForAugmentId';
-export type * from './dto/augment/FindAssetAugmentStatusData';
-export type * from './dto/augment/UpdateAssetAugmentMedia';
-export type * from './dto/auth/AcceptInvitationCompanyParams';
-export type * from './dto/auth/AcceptInvitationCompanyResponse';
-export type * from './dto/auth/AcceptInvitationParams';
-export type * from './dto/auth/AccountRecoveryParams';
-export type * from './dto/auth/InviteUserParams';
-export type * from './dto/auth/PasswordResetParams';
-export type * from './dto/auth/SignInWithPinBody';
-export type * from './dto/auth/SignInWithTokenParams';
-export type * from './dto/auto-slicer/RunScriptParams';
-export type * from './dto/billing/CheckoutSessionParams';
-export type * from './dto/billing/ConsumerBillingInfo';
-export type * from './dto/billing/GetStripeInvoiceParams';
-export type * from './dto/billing/stripeInfo';
-export type * from './dto/billing/StripeInvoiceExtended';
-export type * from './dto/billingCode/BillingCodeResult';
-export type * from './dto/billingCode/GetBillingCodesParams';
-export type * from './dto/billingCode/SearchBillingCodesParams';
-export type * from './dto/catalog/GetCategoriesParams';
-export type * from './dto/catalog/GetProductsParams';
-export type * from './dto/catalog/GetVendorsParams';
-export type * from './dto/clinician/AssignClinicianParams';
-export type * from './dto/clinician/CreateClinicianParams';
-export type * from './dto/clinician/GetCliniciansParams';
-export type * from './dto/company/AddCompanyParams';
-export type * from './dto/company/ClinicalFlowType';
-export type * from './dto/company/CompanyPreferences';
-export type * from './dto/company/CompanyTheme';
-export type * from './dto/company/CompanyWorkbenchWebhook';
-export type * from './dto/company/FindCompaniesParams';
-export type * from './dto/company/GetCompanyByNameParams';
-export type * from './dto/company/UpdateCompanyParams';
-export type * from './dto/email-template/SendEmailTemplateParams';
-export type * from './dto/email-template/UpsertEmailTemplateParams';
-export type * from './dto/email/EmailEvent';
-export type * from './dto/evaluation/ActionEvaluationParams';
-export type * from './dto/evaluation/CreateEvaluationByProductParams';
-export type * from './dto/evaluation/CreateEvaluationParams';
-export type * from './dto/evaluation/EvaluationsStats';
-export type * from './dto/evaluation/EvaluationStatus';
-export type * from './dto/evaluation/GetEvaluationsByPatientParams';
-export type * from './dto/evaluation/GetEvaluationsParams';
-export type * from './dto/evaluation/SearchEvaluationsParams';
-export type * from './dto/evaluation/StartEvaluationByProductParams';
-export type * from './dto/evaluation/UpdateEvaluationParams';
-export type * from './dto/facility/CreateFacilityParams';
-export type * from './dto/facility/SearchFacilityParams';
-export type * from './dto/facility/UpdateFacilityParams';
-export type * from './dto/foot/FootStatus';
-export type * from './dto/foot/SetFootInactive';
-export type * from './dto/foot/UpdateFootFromWorkbenchId';
-export type * from './dto/form/CreateFormTemplateBody';
-export type * from './dto/form/FormTemplateResponse';
-export type * from './dto/form/UpdateFormTemplateBody';
-export type * from './dto/form/UpsertFormSubmissionParams';
-export type * from './dto/invitation/ApprovePatientParams';
-export type * from './dto/invitation/CreateInvitationParams';
-export type * from './dto/invitation/DeleteInvitationsParams';
-export type * from './dto/invitation/FindInvitationsParams';
-export type * from './dto/invitation/SendOtpParams';
-export type * from './dto/invitation/UpdateInvitationsParams';
-export type * from './dto/invitation/VerifyInvitationResponse';
-export type * from './dto/kvs/GetSignalingChannelParams';
-export type * from './dto/kvs/SignalingChannelResponse';
-export type * from './dto/kvs/StreamCompleteParams';
-export type * from './dto/manufacturing/InsertType';
-export type * from './dto/manufacturing/RenderType';
-export type * from './dto/notes/CreateNotes';
-export type * from './dto/notes/Notes';
-export type * from './dto/notes/UpdateNotes';
-export type * from './dto/notification/CreateNotificationMessageParams';
-export type * from './dto/notification/CreateNotificationParams';
-export type * from './dto/notification/EnrollPatientsJobData';
-export type * from './dto/notification/EnrollPatientsParams';
-export type * from './dto/notification/GetNotificationHistoryParams';
-export type * from './dto/notification/GetNotificationsParams';
-export type * from './dto/notification/SendTestParams';
-export type * from './dto/notification/TrackNotificationAction';
-export type * from './dto/notification/UpdateNotificationMessageParams';
-export type * from './dto/order/CreateOrderParams';
-export type * from './dto/order/DeliverOrderParams';
-export type * from './dto/order/GetOrdersByTypeParams';
-export type * from './dto/order/GetOrdersParams';
-export type * from './dto/order/ModifyOrderAuthorizationParams';
-export type * from './dto/order/OrderFormParams';
-export type * from './dto/order/OrdersStats';
-export type * from './dto/order/OrderType';
-export type * from './dto/order/UpdateOrderParams';
-export type * from './dto/PagedParams';
-export type * from './dto/PagedResponse';
-export type * from './dto/patient/CreatePatientParams';
-export type * from './dto/patient/GetPatientsParams';
-export type * from './dto/patient/PatientUserResponse';
-export type * from './dto/patient/SearchPatientsParams';
-export type * from './dto/patient/UpdatePatientParams';
-export type * from './dto/patient/UpsertContactParams';
-export type * from './dto/pdf/PdfContent';
-export type * from './dto/pdf/PdfDoc';
-export type * from './dto/pin/CreatePinDto';
-export type * from './dto/printer/SendGcodeToPrinterParams';
-export type * from './dto/reward/RewardInfo';
-export type * from './dto/roster/DeactivateRosterParams';
-export type * from './dto/roster/DeactivateRosterResponse';
-export type * from './dto/roster/GenerateRosterUploadLinkParams';
-export type * from './dto/roster/ImportRosterParams';
-export type * from './dto/roster/ImportRosterResponse';
-export type * from './dto/roster/RoasterPatientCsvRecord';
-export type * from './dto/scan/CreateMultipartUrls';
-export type * from './dto/scan/SubmitRenderParams';
-export type * from './dto/search/GetSearchParams';
-export type * from './dto/search/SearchResponse';
-export type * from './dto/shipping/GeneratePackingSlip';
-export type * from './dto/shipping/GetLabelsResponse';
-export type * from './dto/shipping/GetLabelTrackingInfo';
-export type * from './dto/shipping/GetRatesResponse';
-export type * from './dto/shipping/GetShipengineLabelsParams';
-export type * from './dto/shipping/GetShipengineLabelsResponse';
-export type * from './dto/shipping/GetShipengineShipmentParams';
-export type * from './dto/shipping/GetShipengineShippingResponse';
-export type * from './dto/shipping/GetShippingLabelLegacyResponse';
-export type * from './dto/shipping/SaveTrackingInfoParams';
-export type * from './dto/shipping/ValidateAddressBody';
-export type * from './dto/tag/UpsertTagParams';
-export type * from './dto/taika/InsoleMeasurements';
-export type * from './dto/terms/AcceptTermsParams';
-export type * from './dto/user/CreateUserParams';
-export type * from './dto/user/GetUsersParams';
-export type * from './dto/user/UpdateUserParams';
-export type * from './dto/workbench/ActionMultipleWorkbenchIdsParams';
-export type * from './dto/workbench/CustomizationResponse';
-export type * from './dto/workbench/DetectionStatusReponse';
-export type * from './dto/workbench/FlattenedWorkbench';
-export type * from './dto/workbench/GenerateWorkbenchPdfParams';
-export type * from './dto/workbench/GetAggregatedParams';
-export type * from './dto/workbench/PatientWorkbenchResponse';
-export type * from './dto/workbench/PrintShippingParams';
-export type * from './dto/workbench/ResetWorkbenchParams';
-export type * from './dto/workbench/SearchWorkbenchParams';
-export type * from './dto/workbench/SubmitOrderParams';
-export type * from './dto/workbench/UpdateInactiveFootBody';
-export type * from './entities/AssetExtended';
-export type * from './entities/CatalogProductExtended';
-export type * from './entities/CatalogProductVariantExtended';
-export type * from './entities/ClinicianExtended';
-export type * from './entities/CompanyExtended';
-export type * from './entities/CompanyPatientExtended';
-export type * from './entities/ContactAddress';
-export type * from './entities/EmailTemplateExtended';
-export type * from './entities/EvaluationExtended';
-export type * from './entities/FacilityExtended';
-export type * from './entities/JobQueueTask';
-export type * from './entities/NotificationExtended';
-export type * from './entities/NotificationHistoryExtended';
-export type * from './entities/NotificationHistoryMessageExtended';
-export type * from './entities/NotificationStats';
-export type * from './entities/OrderExtended';
-export type * from './entities/PatientExtended';
-export type * from './entities/SafeCompany';
-export type * from './entities/SafeUser';
-export type * from './entities/SafeUserExtended';
-export type * from './entities/TagResult';
-export type * from './entities/UserExtended';
-export type * from './entities/WorkbenchExtended';
-export type * from './forms/FieldPrint';
-export type * from './forms/FormField';
-export type * from './forms/FormRule';
-export type * from './forms/FormSchemaTyped';
-export type * from './forms/FormSection';
-export type * from './forms/FormSubmissionExtended';
-export type * from './forms/FormSubmissionTyped';
-export type * from './forms/InvalidFormField';
-export type * from './forms/InvalidFormSection';
+export * from './auth/AuthSession';
+export * from './auth/AuthStatus';
+export * from './auth/AuthUser';
+export * from './config/AppId';
+export * from './config/HikeConfig';
+export * from './config/HikeEnvironment';
+export * from './dto/analytics/HourlyOptions';
+export * from './dto/appointment/GenerateAppointmentsUploadLinkParams';
+export * from './dto/appointment/ImportAppointmentsParams';
+export * from './dto/appointment/ImportAppointmentsResponse';
+export * from './dto/appointment/ImportPrescriptionsResponse';
+export * from './dto/asset/GetAssetsParams';
+export * from './dto/augment/AssetAugmentData';
+export * from './dto/augment/AssetAugmentMediaData';
+export * from './dto/augment/AssetAugmentMediaResult';
+export * from './dto/augment/AssetAugmentResult';
+export * from './dto/augment/AssetAugmentStatus';
+export * from './dto/augment/AssetAugmentStatusForAugmentId';
+export * from './dto/augment/FindAssetAugmentStatusData';
+export * from './dto/augment/UpdateAssetAugmentMedia';
+export * from './dto/auth/AcceptInvitationCompanyParams';
+export * from './dto/auth/AcceptInvitationCompanyResponse';
+export * from './dto/auth/AcceptInvitationParams';
+export * from './dto/auth/AccountRecoveryParams';
+export * from './dto/auth/InviteUserParams';
+export * from './dto/auth/PasswordResetParams';
+export * from './dto/auth/SignInWithPinBody';
+export * from './dto/auth/SignInWithTokenParams';
+export * from './dto/auto-slicer/RunScriptParams';
+export * from './dto/billing/CheckoutSessionParams';
+export * from './dto/billing/ConsumerBillingInfo';
+export * from './dto/billing/GetStripeInvoiceParams';
+export * from './dto/billing/stripeInfo';
+export * from './dto/billing/StripeInvoiceExtended';
+export * from './dto/billingCode/BillingCodeResult';
+export * from './dto/billingCode/GetBillingCodesParams';
+export * from './dto/billingCode/SearchBillingCodesParams';
+export * from './dto/catalog/GetCategoriesParams';
+export * from './dto/catalog/GetProductsParams';
+export * from './dto/catalog/GetVendorsParams';
+export * from './dto/clinician/AssignClinicianParams';
+export * from './dto/clinician/CreateClinicianParams';
+export * from './dto/clinician/GetCliniciansParams';
+export * from './dto/company/AddCompanyParams';
+export * from './dto/company/ClinicalFlowType';
+export * from './dto/company/CompanyPortal';
+export * from './dto/company/CompanyPreferences';
+export * from './dto/company/CompanyTheme';
+export * from './dto/company/CompanyWorkbenchWebhook';
+export * from './dto/company/FindCompaniesParams';
+export * from './dto/company/GetCompanyByNameParams';
+export * from './dto/company/UpdateCompanyParams';
+export * from './dto/email-template/SendEmailTemplateParams';
+export * from './dto/email-template/UpsertEmailTemplateParams';
+export * from './dto/email/EmailEvent';
+export * from './dto/evaluation/ActionEvaluationParams';
+export * from './dto/evaluation/CreateEvaluationByProductParams';
+export * from './dto/evaluation/CreateEvaluationParams';
+export * from './dto/evaluation/EvaluationsStats';
+export * from './dto/evaluation/EvaluationStatus';
+export * from './dto/evaluation/GetEvaluationsByPatientParams';
+export * from './dto/evaluation/GetEvaluationsParams';
+export * from './dto/evaluation/SearchEvaluationsParams';
+export * from './dto/evaluation/StartEvaluationByProductParams';
+export * from './dto/evaluation/UpdateEvaluationParams';
+export * from './dto/facility/CreateFacilityParams';
+export * from './dto/facility/SearchFacilityParams';
+export * from './dto/facility/UpdateFacilityParams';
+export * from './dto/foot/FootStatus';
+export * from './dto/foot/SetFootInactive';
+export * from './dto/foot/UpdateFootFromWorkbenchId';
+export * from './dto/form/CreateFormTemplateBody';
+export * from './dto/form/FormTemplateResponse';
+export * from './dto/form/UpdateFormTemplateBody';
+export * from './dto/form/UpsertFormSubmissionParams';
+export * from './dto/invitation/ApprovePatientParams';
+export * from './dto/invitation/CreateInvitationParams';
+export * from './dto/invitation/DeleteInvitationsParams';
+export * from './dto/invitation/FindInvitationsParams';
+export * from './dto/invitation/SendOtpParams';
+export * from './dto/invitation/UpdateInvitationsParams';
+export * from './dto/invitation/VerifyInvitationResponse';
+export * from './dto/kvs/GetSignalingChannelParams';
+export * from './dto/kvs/SignalingChannelResponse';
+export * from './dto/kvs/StreamCompleteParams';
+export * from './dto/manufacturing/InsertType';
+export * from './dto/manufacturing/RenderType';
+export * from './dto/notes/CreateNotes';
+export * from './dto/notes/Notes';
+export * from './dto/notes/UpdateNotes';
+export * from './dto/notification/CreateNotificationMessageParams';
+export * from './dto/notification/CreateNotificationParams';
+export * from './dto/notification/EnrollPatientsJobData';
+export * from './dto/notification/EnrollPatientsParams';
+export * from './dto/notification/GetNotificationHistoryParams';
+export * from './dto/notification/GetNotificationsParams';
+export * from './dto/notification/SendTestParams';
+export * from './dto/notification/TrackNotificationAction';
+export * from './dto/notification/UpdateNotificationMessageParams';
+export * from './dto/order/CreateOrderParams';
+export * from './dto/order/DeliverOrderParams';
+export * from './dto/order/GetOrdersByTypeParams';
+export * from './dto/order/GetOrdersParams';
+export * from './dto/order/ModifyOrderAuthorizationParams';
+export * from './dto/order/OrderFormParams';
+export * from './dto/order/OrdersStats';
+export * from './dto/order/OrderType';
+export * from './dto/order/UpdateOrderParams';
+export * from './dto/PagedParams';
+export * from './dto/PagedResponse';
+export * from './dto/patient/CreatePatientParams';
+export * from './dto/patient/GetPatientsParams';
+export * from './dto/patient/PatientUserResponse';
+export * from './dto/patient/SearchPatientsParams';
+export * from './dto/patient/UpdatePatientParams';
+export * from './dto/patient/UpsertContactParams';
+export * from './dto/pdf/PdfContent';
+export * from './dto/pdf/PdfDoc';
+export * from './dto/pin/CreatePinDto';
+export * from './dto/printer/SendGcodeToPrinterParams';
+export * from './dto/reward/RewardInfo';
+export * from './dto/roster/DeactivateRosterParams';
+export * from './dto/roster/DeactivateRosterResponse';
+export * from './dto/roster/GenerateRosterUploadLinkParams';
+export * from './dto/roster/ImportRosterParams';
+export * from './dto/roster/ImportRosterResponse';
+export * from './dto/roster/RoasterPatientCsvRecord';
+export * from './dto/scan/CreateMultipartUrls';
+export * from './dto/scan/SubmitRenderParams';
+export * from './dto/search/GetSearchParams';
+export * from './dto/search/SearchResponse';
+export * from './dto/shipping/GeneratePackingSlip';
+export * from './dto/shipping/GetLabelsResponse';
+export * from './dto/shipping/GetLabelTrackingInfo';
+export * from './dto/shipping/GetRatesResponse';
+export * from './dto/shipping/GetShipengineLabelsParams';
+export * from './dto/shipping/GetShipengineLabelsResponse';
+export * from './dto/shipping/GetShipengineShipmentParams';
+export * from './dto/shipping/GetShipengineShippingResponse';
+export * from './dto/shipping/GetShippingLabelLegacyResponse';
+export * from './dto/shipping/SaveTrackingInfoParams';
+export * from './dto/shipping/ValidateAddressBody';
+export * from './dto/tag/UpsertTagParams';
+export * from './dto/taika/InsoleMeasurements';
+export * from './dto/terms/AcceptTermsParams';
+export * from './dto/user/CreateUserParams';
+export * from './dto/user/GetUsersParams';
+export * from './dto/user/UpdateUserParams';
+export * from './dto/workbench/ActionMultipleWorkbenchIdsParams';
+export * from './dto/workbench/CustomizationResponse';
+export * from './dto/workbench/DetectionStatusReponse';
+export * from './dto/workbench/FlattenedWorkbench';
+export * from './dto/workbench/GenerateWorkbenchPdfParams';
+export * from './dto/workbench/GetAggregatedParams';
+export * from './dto/workbench/PatientWorkbenchResponse';
+export * from './dto/workbench/PrintShippingParams';
+export * from './dto/workbench/ResetWorkbenchParams';
+export * from './dto/workbench/SearchWorkbenchParams';
+export * from './dto/workbench/SubmitOrderParams';
+export * from './dto/workbench/UpdateInactiveFootBody';
+export * from './entities/AssetExtended';
+export * from './entities/CatalogProductExtended';
+export * from './entities/CatalogProductVariantExtended';
+export * from './entities/ClinicianExtended';
+export * from './entities/CompanyExtended';
+export * from './entities/CompanyPatientExtended';
+export * from './entities/ContactAddress';
+export * from './entities/EmailTemplateExtended';
+export * from './entities/EvaluationExtended';
+export * from './entities/FacilityExtended';
+export * from './entities/JobQueueTask';
+export * from './entities/NotificationExtended';
+export * from './entities/NotificationHistoryExtended';
+export * from './entities/NotificationHistoryMessageExtended';
+export * from './entities/NotificationStats';
+export * from './entities/OrderExtended';
+export * from './entities/PatientExtended';
+export * from './entities/SafeCompany';
+export * from './entities/SafeUser';
+export * from './entities/SafeUserExtended';
+export * from './entities/TagResult';
+export * from './entities/UserExtended';
+export * from './entities/WorkbenchExtended';
+export * from './forms/FieldPrint';
+export * from './forms/FormField';
+export * from './forms/FormRule';
+export * from './forms/FormSchemaTyped';
+export * from './forms/FormSection';
+export * from './forms/FormSubmissionExtended';
+export * from './forms/FormSubmissionTyped';
+export * from './forms/InvalidFormField';
+export * from './forms/InvalidFormSection';
+export * from './guards/isDefined';
+export * from './guards/isEmpty';
+export * from './guards/isFieldPrint';
+export * from './guards/isFormFieldValue';
+export * from './guards/isFormSubmissionData';
+export * from './guards/isFulfilled';
+export * from './guards/isMinimumRole';
+export * from './guards/isNotEmpty';
+export * from './guards/isPlainObject';
+export * from './guards/isPrintableField';
+export * from './guards/isString';
+export * from './guards/isTruthy';
