+++ ui/src/hooks/clinician/useCreateClinician.ts
@@ -2,13 +2,9 @@ import { HikeError, createClinician } from '@hike/services';
 import { Clinician, CreateClinicianParams } from '@hike/types';
 import { UseMutationOptions, useMutation } from '@tanstack/react-query';
 
-interface CreateClinicianOptions {
-  params: CreateClinicianParams;
-}
-
-export const useCreateClinician = (options?: UseMutationOptions<Clinician, HikeError<null>, CreateClinicianOptions>) =>
+export const useCreateClinician = (options?: UseMutationOptions<Clinician, HikeError<null>, CreateClinicianParams>) =>
   useMutation({
     mutationKey: ['createClinician'],
-    mutationFn: async ({ params }) => await createClinician(params),
+    mutationFn: async (params) => await createClinician(params),
     ...options
   });
