+++ ui/src/hooks/company/useAddCompany.ts
@@ -2,13 +2,9 @@ import { HikeError, addCompany } from '@hike/services';
 import { AddCompanyParams, CompanyExtended } from '@hike/types';
 import { UseMutationOptions, useMutation } from '@tanstack/react-query';
 
-interface AddCompanyOptions {
-  body: AddCompanyParams;
-}
-
-export const useAddCompany = (options?: UseMutationOptions<CompanyExtended, HikeError<null>, AddCompanyOptions>) =>
+export const useAddCompany = (options?: UseMutationOptions<CompanyExtended, HikeError<null>, AddCompanyParams>) =>
   useMutation({
     mutationKey: ['addCompany'],
-    mutationFn: async ({ body }) => await addCompany(body),
+    mutationFn: async (params) => await addCompany(params),
     ...options
   });
