+++ ui/src/hooks/facilities/useCreateFacility.ts
@@ -2,15 +2,11 @@ import { HikeError, createFacility } from '@hike/services';
 import { CreateFacilityParams, FacilityExtended } from '@hike/types';
 import { UseMutationOptions, useMutation } from '@tanstack/react-query';
 
-interface CreateFacilityContext {
-  body: CreateFacilityParams;
-}
-
 export const useCreateFacility = (
-  options?: UseMutationOptions<FacilityExtended, HikeError<null>, CreateFacilityContext>
+  options?: UseMutationOptions<FacilityExtended, HikeError<null>, CreateFacilityParams>
 ) =>
   useMutation({
     mutationKey: ['createFacility'],
-    mutationFn: async ({ body }) => await createFacility(body),
+    mutationFn: async (params) => await createFacility(params),
     ...options
   });
