+++ utils/src/index.ts
@@ -10,23 +10,12 @@ export * from './converters/parseCityStateZip';
 export * from './converters/parseClinicianName';
 export * from './converters/parseDate';
 export * from './converters/parsePersonName';
-export * from './converters/toAuthUser';
 export * from './converters/toBoolean';
 export * from './converters/toJobQueueTask';
 export * from './converters/toTitleCase';
 export * from './converters/toURL';
 export * from './converters/toUTC';
 export * from './geography';
-export * from './guards/isAuthUser';
-export * from './guards/isDefined';
-export * from './guards/isFieldPrint';
-export * from './guards/isFormFieldValue';
-export * from './guards/isFormSubmissionData';
-export * from './guards/isFulfilled';
-export * from './guards/isNotEmpty';
-export * from './guards/isPrintableField';
-export * from './guards/isString';
-export * from './guards/isTruthy';
 export * from './helpers/addHeaders';
 export * from './helpers/appUtils';
 export * from './helpers/colorMapper';
@@ -39,8 +28,6 @@ export * from './helpers/formAddressUtils';
 export * from './helpers/formSchemaUtils';
 export * from './helpers/getLocalizedDateFormat';
 export * from './helpers/httpUtils';
-export * from './helpers/isEmpty';
-export * from './helpers/isMinimumRole';
 export * from './helpers/localeUtils';
 export * from './helpers/postMessageToSlack';
 export * from './helpers/queryStringUtils';
